Changed money type from double to BigDecimal. Added setting to gradle task, so that resources folder would be copied to build directory/libs

+++ src/main/java/lma/task1/dto/SummaryRow.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -21,7 +22,7 @@ public class SummaryRow {
 
     private Integer transactionCount;
 
-    private Double debt;
+    private BigDecimal debt;
 
     private PeriodEnum period;
